+++ live114-no_shake/CardLeap/Resource/Map/library/JPSThumbnailAnnotation/JPSThumnail.m
@@ -12,7 +12,7 @@
 
 - (NSString *)description
 {
-    return [NSString stringWithFormat:@"tag : %d",self.tag];
+    return [NSString stringWithFormat:@"tag : %ld",(long)self.tag];
 }
 
 @end
